aguante/requirements.txt
Django==1.7.1
Pillow==2.6.1
djangorestframework==2.4.4
django-grappelli==2.6.3

ipdb
